+++ libs/util/code-gen-utils/src/lib/index.ts
@@ -1,2 +1,3 @@
 export * from "./enum-builder";
 export * from "./files";
+export * from "./parse";
