+++ libs/util/code-gen-utils/src/lib/parse/errors/ParseError.ts
@@ -1,0 +1,5 @@
+export class ParseError extends SyntaxError {
+  constructor(message: string, source: string) {
+    super(`${message}\nSource:\n${source}`);
+  }
+}
