+++ libs/util/code-gen-utils/src/lib/parse/partial-parser/index.ts
@@ -1,0 +1,1 @@
+export * from "./partial-parser";
